feat: Implement Search and clear search functionality
- Added `MySearchBar` composable for searching anime.
- Integrated search functionality within `HomeViewModel` to fetch anime based on user input.
- Added clear search functionality to the view model.
- Updated `HomeScreen` to display search results and manage the search state.
- Updated the back handler to clear the search state.
- Updated MainActivity for edge to edge behavior.
