+++ app/src/main/java/com/example/seekhoandoridassignment/presntation/common/SearchUI.kt
@@ -1,0 +1,95 @@
+package com.example.seekhoandoridassignment.presntation.common
+
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.verticalScroll
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Close
+import androidx.compose.material.icons.filled.Search
+import androidx.compose.material.icons.filled.Star
+import androidx.compose.material3.DockedSearchBar
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.ListItem
+import androidx.compose.material3.ListItemDefaults
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.SearchBar
+import androidx.compose.material3.SearchBarDefaults
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.semantics.isTraversalGroup
+import androidx.compose.ui.semantics.semantics
+import androidx.compose.ui.semantics.traversalIndex
+import androidx.compose.ui.unit.dp
+import com.skydoves.cloudy.cloudy
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun MySearchBar( search: (searchQuery: String) -> Unit) {
+    var textFieldState by remember { mutableStateOf("") }
+    var expanded by rememberSaveable { mutableStateOf(false) }
+
+    Box(Modifier.fillMaxSize().semantics { isTraversalGroup = true }) {
+        DockedSearchBar(
+            modifier = Modifier
+                .align(Alignment.TopCenter)
+                .fillMaxWidth()
+                .semantics { traversalIndex = 0f },
+            query = textFieldState,
+            onQueryChange = { textFieldState = it },
+            onSearch = { search(textFieldState) ;expanded = false },
+            active = expanded,
+            onActiveChange = { expanded = it },
+            placeholder = { Text(text = " search text", color = Color.Black) },
+            leadingIcon = { Icon(Icons.Filled.Search, contentDescription = "Search Icon") },
+            trailingIcon = {
+                if (textFieldState.isNotEmpty()) {
+                    IconButton(onClick = { textFieldState = "" }) {
+                        Icon(Icons.Filled.Close, contentDescription = "Clear text")
+                    }
+                }
+            },
+            shape = MaterialTheme.shapes.small,
+            colors = SearchBarDefaults.colors(
+                containerColor = Color.White,
+            ),
+        )
+        {
+            Column(Modifier.verticalScroll(rememberScrollState())) {
+                repeat(4) { idx ->
+                    val resultText = "Suggestion $idx"
+                    ListItem(
+                        headlineContent = { Text(resultText) },
+                        supportingContent = { Text("Additional info") },
+                        leadingContent = { Icon(Icons.Filled.Star, contentDescription = null) },
+                        colors = ListItemDefaults.colors(containerColor = Color.Transparent),
+                        modifier = Modifier
+                            .clickable {
+                                textFieldState = resultText
+                                expanded = false
+                            }
+                            .fillMaxWidth()
+                            .padding(horizontal = 16.dp, vertical = 4.dp)
+                    )
+                }
+            }
+        }
+
+
+    }
+}
