+++ app/src/main/java/com/example/seekhoandoridassignment/presntation/screens/HomeScreen.kt
@@ -18,21 +18,9 @@ import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.grid.GridCells
 import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
 import androidx.compose.foundation.lazy.grid.rememberLazyGridState
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.verticalScroll
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Close
-import androidx.compose.material.icons.filled.Search
-import androidx.compose.material.icons.filled.Star
 import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.ListItem
-import androidx.compose.material3.ListItemDefaults
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.SearchBar
-import androidx.compose.material3.SearchBarDefaults
 import androidx.compose.material3.Text
 import androidx.compose.runtime.*
 import androidx.compose.runtime.Composable
@@ -41,7 +29,6 @@ import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -49,9 +36,6 @@ import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.semantics.isTraversalGroup
-import androidx.compose.ui.semantics.semantics
-import androidx.compose.ui.semantics.traversalIndex
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
@@ -63,9 +47,9 @@ import coil.compose.SubcomposeAsyncImage
 import coil.request.CachePolicy
 import com.example.seekhoandoridassignment.data.dto.Anime
 import com.example.seekhoandoridassignment.data.dto.AnimeDetailsDto
+import com.example.seekhoandoridassignment.presntation.common.MySearchBar
 import com.example.seekhoandoridassignment.presntation.viewmodels.HomeViewModel
 import com.example.seekhoandoridassignment.uitl.ApiState
-import com.skydoves.cloudy.cloudy
 import kotlinx.serialization.Serializable
 import org.koin.androidx.compose.koinViewModel
 
@@ -75,7 +59,7 @@ fun HomeScreen(navController: NavController) {
     val viewModel: HomeViewModel = koinViewModel()
     val animeListState = viewModel.animeList.collectAsLazyPagingItems()
     val animeDetailState = viewModel.animeDetailState.collectAsState()
-    val animeSearchList = viewModel.animeSearch.collectAsState().value
+    var animeSearchList = viewModel.animeSearch.collectAsState().value
     var animeDetails = AnimeDetailsDto(title = "title", trailer = null, plot = "", genres = null, mainCast = null, noOfEpisodes = null, imageUrl = "")
     var dominantColor = viewModel.color.collectAsState()
     val detailview = remember { mutableStateOf(false)}
@@ -101,7 +85,6 @@ fun HomeScreen(navController: NavController) {
                 lastVisibleItemIndex = visibleItemIndex
                 lastScrollOffset = scrollOffset
 
-
             }
     }
 
@@ -111,6 +94,9 @@ fun HomeScreen(navController: NavController) {
     }
     BackHandler(enabled = detailview.value) {
         detailview.value = false
+    }
+    BackHandler(enabled = true) {
+      viewModel.clearSearch()
     }
     LaunchedEffect(Unit) {
         if (animeListState.loadState.refresh is LoadState.Loading) {
@@ -178,61 +164,7 @@ fun HomeScreen(navController: NavController) {
 
 }
 
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun MySearchBar( search: (searchQuery: String) -> Unit) {
-    var textFieldState by remember { mutableStateOf("") }
-    var expanded by rememberSaveable { mutableStateOf(false) }
 
-    Box(Modifier.fillMaxSize().semantics { isTraversalGroup = true }) {
-        SearchBar(
-            modifier = Modifier
-                .align(Alignment.TopCenter)
-                .cloudy()
-                .semantics { traversalIndex = 0f },
-            query = textFieldState,
-            onQueryChange = { textFieldState = it },
-            onSearch = { search(textFieldState) },
-            active = expanded,
-            onActiveChange = { expanded = it },
-            placeholder = { Text(text = "Hinted search text", color = Color.Black) },
-            leadingIcon = { Icon(Icons.Filled.Search, contentDescription = "Search Icon") },
-            trailingIcon = {
-                if (textFieldState.isNotEmpty()) {
-                    IconButton(onClick = { textFieldState = "" }) {
-                        Icon(Icons.Filled.Close, contentDescription = "Clear text")
-                    }
-                }
-            },
-            shape = MaterialTheme.shapes.small,
-            colors = SearchBarDefaults.colors(
-                containerColor = Color.Gray.copy(0.6f),
-            ),
-        )
-        {
-            Column(Modifier.verticalScroll(rememberScrollState())) {
-                repeat(4) { idx ->
-                    val resultText = "Suggestion $idx"
-                    ListItem(
-                        headlineContent = { Text(resultText) },
-                        supportingContent = { Text("Additional info") },
-                        leadingContent = { Icon(Icons.Filled.Star, contentDescription = null) },
-                        colors = ListItemDefaults.colors(containerColor = Color.Transparent),
-                        modifier = Modifier
-                            .clickable {
-                                textFieldState = resultText
-                                expanded = false
-                            }
-                            .fillMaxWidth()
-                            .padding(horizontal = 16.dp, vertical = 4.dp)
-                    )
-                }
-            }
-        }
-
-
-    }
-}
